rename typo rsna_penumonia.py to rsna_pneumonia.py and update import statements in rsna_ft.py. renamed test to env_test folder
